test: Split models.utils unittests
